pyproject.toml
[tool.poetry]
name = "weather-app"
version = "0.1.0"
description = ""
authors = ["Cur1yB <alexandr.ignatev1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
requests = "^2.32.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
